+++ src/core_modules/capture-core/components/D2Form/SectionDescriptionBox.types.js
@@ -1,0 +1,13 @@
+// @flow
+export type Props = {|
+    description: string,
+|};
+
+export type PlainProps = {|
+    classes: {
+        descriptionBox: string,
+        icon: string,
+        description: string,
+    },
+    ...Props,
+|};