+++ src/core_modules/capture-core/components/DataEntry/common/dataEntryHasChanges.js
@@ -1,9 +1,15 @@
 // @flow
 
-export const getDataEntryHasChanges = (state: ReduxState, key: string) => {
+export const dataEntryHasChanges = (state: ReduxState, key: string): boolean => {
     const reduced = Object.keys(state.formsSectionsFieldsUI)
         .filter(formSectionUI => formSectionUI.startsWith(key))
-        .reduce((accElementsUI, sectionKey) => [...accElementsUI, ...Object.keys(state.formsSectionsFieldsUI[sectionKey]).map(elementKey => state.formsSectionsFieldsUI[sectionKey][elementKey])], []);
+        .reduce((accElementsUI, sectionKey) =>
+            [
+                ...accElementsUI,
+                ...Object.keys(state.formsSectionsFieldsUI[sectionKey])
+                    .map(elementKey => state.formsSectionsFieldsUI[sectionKey][elementKey]),
+            ]
+        , []);
 
     const formIsModified = reduced.some(element => element.modified);