+++ src/core_modules/capture-core/components/Dialogs/ConfirmDialog.component.js
@@ -1,5 +1,6 @@
 // @flow
 import React, { Component } from 'react';
+import { withStyles } from '@material-ui/core';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
@@ -17,6 +18,9 @@ type Props = {
     onConfirm: () => void,
 };
 
+const StyledDialog = withStyles({ root: { zIndex: 3000 } })(Dialog);
+const StyledDialogActions = withStyles({ root: { margin: '8px 8px 12px 0' } })(DialogActions);
+
 export class ConfirmDialog extends Component<Props> {
     render() {
         const {
@@ -30,7 +34,7 @@ export class ConfirmDialog extends Component<Props> {
         } = this.props;
 
         return (
-            <Dialog
+            <StyledDialog
                 open={open}
                 onClose={onCancel}
             >
@@ -40,15 +44,15 @@ export class ConfirmDialog extends Component<Props> {
                         {text}
                     </DialogContentText>
                 </DialogContent>
-                <DialogActions>
+                <StyledDialogActions>
                     <Button onClick={onCancel} secondary>
                         {cancelText}
                     </Button>
                     <Button onClick={onConfirm} primary>
                         {confirmText}
                     </Button>
-                </DialogActions>
-            </Dialog>
+                </StyledDialogActions>
+            </StyledDialog>
         );
     }
 }