+++ src/core_modules/capture-core/components/Pages/ViewEvent/EventDetailsSection/ViewEventDataEntry/ViewEventDataEntry.component.js
@@ -151,7 +151,7 @@ const buildGeometrySettingsFn = () => ({
         });
 
         return createComponentProps(props, {
-            label: i18n.t('Coordinate'),
+            label: 'Coordinate',
             valueConverter: value => pointDataElement.convertValue(value, valueConvertFn),
         });
     },
@@ -172,7 +172,7 @@ const buildCompleteFieldSettingsFn = () => {
     const completeSettings = {
         getComponent: () => viewModeComponent,
         getComponentProps: (props: Object) => createComponentProps(props, {
-            label: i18n.t('Event completed'),
+            label: 'Event completed',
             id: dataElement.id,
             valueConverter: value => dataElement.convertValue(value, valueConvertFn),
         }),