+++ src/core_modules/capture-core/components/Pages/ViewEvent/ViewEventPage.container.js
@@ -1,15 +1,17 @@
 // @flow
 import { connect } from 'react-redux';
 import { ViewEventPageComponent } from './ViewEventPage.component';
-import { getDataEntryHasChanges } from '../../DataEntry/common/dataEntryHasChanges';
+import { dataEntryHasChanges } from '../../DataEntry/common/dataEntryHasChanges';
 import { withLoadingIndicator, withErrorMessageHandler } from '../../../HOC';
 
 const mapStateToProps = (state: ReduxState) => {
     const eventDetailsSection = state.viewEventPage.eventDetailsSection || {};
     const isUserInteractionInProgress =
-        state.currentSelections.complete &&
-        eventDetailsSection.showEditEvent &&
-        getDataEntryHasChanges(state, 'singleEvent-editEvent');
+      (state.currentSelections.complete && eventDetailsSection.showEditEvent)
+          ?
+          dataEntryHasChanges(state, 'singleEvent-editEvent')
+          :
+          false;
     return {
         error: state.activePage.viewEventLoadError && state.activePage.viewEventLoadError.error,
         ready: !state.activePage.lockedSelectorLoads,