+++ src/core_modules/capture-core/metaData/RenderFoundation/Section.js
@@ -15,6 +15,7 @@ export class Section {
 
     _id: string;
     _name: string;
+    _displayDescription: string;
     _open: boolean;
     _visible: boolean;
     _collapsible: boolean;
@@ -42,6 +43,13 @@ export class Section {
     }
     get name(): string {
         return this._name;
+    }
+
+    set displayDescription(description: string) {
+        this._displayDescription = description;
+    }
+    get displayDescription(): string {
+        return this._displayDescription;
     }
 
     set open(open: ?boolean) {