+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/enrollment/DataElementFactory.js
@@ -85,6 +85,7 @@ export class DataElementFactory {
                         .get(
                             'trackedEntityInstances',
                             {
+                                program: contextProps.programId,
                                 ou: orgUnitId,
                                 filter: `${dataElement.id}:EQ:${serverValue}`,
                             },
@@ -94,6 +95,7 @@ export class DataElementFactory {
                         .get(
                             'trackedEntityInstances',
                             {
+                                program: contextProps.programId,
                                 ouMode: 'ACCESSIBLE',
                                 filter: `${dataElement.id}:EQ:${serverValue}`,
                             },