+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/programStage/ProgramStageFactory.js
@@ -50,6 +50,7 @@ export class ProgramStageFactory {
         const section = new Section((o) => {
             o.id = sectionSpecs.id;
             o.name = sectionSpecs.displayName;
+            o.displayDescription = sectionSpecs.displayDescription;
         });
 
         if (sectionSpecs.dataElements) {
@@ -168,6 +169,7 @@ export class ProgramStageFactory {
                 stageForm.addSection(await this._buildSection(cachedProgramStageDataElementsAsObject, {
                     id: section.id,
                     displayName: section.displayName,
+                    displayDescription: section.displayDescription,
                     dataElements: section.dataElements,
                 }));
             });